+++ app/src/main/java/lavender/client/android/data/grpc/GrpcReconnectStrategy.kt
@@ -1,7 +1,10 @@
 package lavender.client.android.data.grpc
 
 import android.util.Log
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 import kotlinx.coroutines.flow.MutableStateFlow
 
 /**