+++ app/src/main/java/lavender/client/android/ui/chat/message/ChatSelectionDelegate.kt
@@ -180,8 +180,8 @@ class ChatSelectionDelegate(
                         hideSelectionToolbar()
                     }
                 )
-                sheet.loadChats(f)
                 sheet.show()
+                sheet.loadChats(f)
             }
         }
     }